Moved TestInteractor into separate file.
